Tests/LibWeb/Text/input/URL/strip_trailing_spaces_from_opaque_path.html

<script src="../include.js"></script>
<script>
    test(() => {
        let url = new URL("scheme:foobar                 ?well=hello");
        println(`pathname => '${url.pathname}'`);
        url.search = "";
        println(`pathname => '${url.pathname}'`);

        url = new URL("scheme:baz #friends");
        println(`pathname => '${url.pathname}'`);
        url.hash = "";
        println(`pathname => '${url.pathname}'`);
    });
</script>
