Tests/LibWeb/Text/expected/URL/strip_trailing_spaces_from_opaque_path.txt

pathname => 'foobar                 '
pathname => 'foobar'
pathname => 'baz '
pathname => 'baz'
